Remove MediatR from the AdventOfCodeSkill
