+++ Guppi.Application/DependencyInjection.cs
@@ -1,16 +1,18 @@
 using System.Reflection;
+using Guppi.Application.Services;
 using MediatR;
 using Microsoft.Extensions.DependencyInjection;
 
-namespace Guppi.Application
-{
-    public static class DependencyInjection
-    {
-        public static IServiceCollection AddApplication(this IServiceCollection services)
-        {
-            services.AddMediatR(Assembly.GetExecutingAssembly());
+namespace Guppi.Application;
 
-            return services;
-        }
+public static class DependencyInjection
+{
+    public static IServiceCollection AddApplication(this IServiceCollection services)
+    {
+        services
+            .AddMediatR(Assembly.GetExecutingAssembly())
+            .AddTransient<IAdventOfCodeService, AdventOfCodeService>();
+
+        return services;
     }
 }
